+++ lib/keynote/presenter.rb
@@ -113,6 +113,13 @@ module Keynote
       end
     end
 
+    # @private
+    # We have to make a logger method available so that ActionView::Template
+    # can safely treat a presenter as a view object.
+    def logger
+      Rails.logger
+    end
+
     private
 
     # We have to explicitly proxy `#capture` because ActiveSupport creates a