+++ CHANGELOG.md
@@ -1,10 +1,17 @@
-## v0.2.0pre1
+## v0.2.0
+* Add `Keynote::Inline`, a module that presenters can extend to enable inline
+  templating in any language supported by Rails.
+* Stop including the `Keynote::Rumble` module in presenters by default;
+  presenters that want to use the `build_html` method must explicitly include
+  it.
+* Including `Keynote::Rumble` now provides a `use_html_5_tags` class method
+  that adds a more complete set of Rumble tag methods to the class.
+* Add `object_names` class method to presenters, returning an array of the
+  symbols that have been passed into the `presents` method.
 * Add an implementation of the `present`/`k` method that's available in test
   cases. Update test generators accordingly.
 * Update minitest-rails integration to be compatible with the newest
   version of minitest-rails (on Rails 3.0, 3.1, 3.2, and 4.0).
-* Add `use_html_5_tags` class method to `Keynote::Presenter`. This adds a more
-  complete set of Rumble tag methods to the class.
 
 ## v0.1.3
 * Add block form of `present`. If you pass a block into a `present` call, it